Move retry config to `nextflow.retryPolicy`
