+++ modules/nextflow/src/main/groovy/nextflow/util/RetryConfig.groovy
@@ -15,32 +15,31 @@
  *
  */
 
-package nextflow.processor
+package nextflow.util
 
 import groovy.transform.CompileStatic
 import groovy.transform.EqualsAndHashCode
 import groovy.transform.ToString
-import nextflow.util.Duration
 
 /**
- * Models retry policy configuration for publishing outputs
+ * Models retry policy configuration
  *
- * @author Ben Sherman <bentshermann@gmail.com>
+ * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
 @ToString(includePackage = false, includeNames = true)
 @EqualsAndHashCode
 @CompileStatic
-class PublishRetryConfig {
+class RetryConfig {
     Duration delay = Duration.of('350ms')
     Duration maxDelay = Duration.of('90s')
     int maxAttempts = 5
     double jitter = 0.25
 
-    PublishRetryConfig() {
+    RetryConfig() {
         this(Collections.emptyMap())
     }
 
-    PublishRetryConfig(Map config) {
+    RetryConfig(Map config) {
         if( config.delay )
             delay = config.delay as Duration
         if( config.maxDelay )
